+++ packages/core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @trainheroic-unofficial/core
 
+## 0.5.0
+
+### Patch Changes
+
+- @trainheroic-unofficial/dto@0.5.0
+- @trainheroic-unofficial/js@0.5.0
+
 ## 0.4.2
 
 ### Patch Changes